requirements.txt
pandas
opencv-python
Pillow
detecto